character_creation_package/pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "character_creation"
version = "0.1.0"
description = "Data-driven character creation system for WorldSeedProject."
authors = [
    { name = "WorldSeedProject Team" }
]
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
dependencies = [
    "pyyaml>=6.0",
    "textual>=0.60",
    "PySide6>=6.7"
]

[project.scripts]
worldseed-wizard = "scripts.create_character:main"
worldseed-tui = "scripts.run_tui:main"
worldseed-validate = "scripts.validate_data:main"
worldseed-npc-demo = "scripts.generate_npcs:main"
worldseed-xp-demo = "scripts.grant_xp:main"
worldseed-regen-demo = "scripts.sim_regen:main"
worldseed-effects-demo = "scripts.sim_status_effects:main"
worldseed-save-demo = "scripts.demo_save_load:main"
worldseed-aurora = "scripts.run_qt_aurora:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "black",
    "ruff",
    "pre-commit",
    "watchfiles>=0.21"
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.setuptools.packages.find]
where = ["."]
include = ["character_creation*", "scripts*"]

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
ignore_missing_imports = true
